+++ src/Polynomials/QuarticPolynomial.cpp
@@ -5,8 +5,8 @@
 
 using namespace Eigen;
 
-QuarticPolynomial::QuarticPolynomial(double xs, double vxs, double axs,
-        double vxe, double axe, double t):
+QuarticPolynomial::QuarticPolynomial(float xs, float vxs, float axs,
+        float vxe, float axe, float t):
         a0(xs), a1(vxs) {
     a2 = axs / 2.0;
     Matrix2d A;
@@ -19,18 +19,18 @@ QuarticPolynomial::QuarticPolynomial(double xs, double vxs, double axs,
     a4 = x[1];
 }
 
-fixp_s QuarticPolynomial::calc_point(double t) {
+fixp_s QuarticPolynomial::calc_point(float t) {
     return a0 + a1 * t + a2 * pow(t, 2) + a3 * pow(t, 3) + a4 * pow(t, 4);
 }
 
-fixp_s_d QuarticPolynomial::calc_first_derivative(double t) {
+fixp_s_d QuarticPolynomial::calc_first_derivative(float t) {
     return a1 + 2 * a2 * t + 3 * a3 * pow(t, 2) + 4 * a4 * pow(t, 3);
 }
 
-fixp_s_dd QuarticPolynomial::calc_second_derivative(double t) {
+fixp_s_dd QuarticPolynomial::calc_second_derivative(float t) {
     return 2 * a2 + 6 * a3 * t + 12 * a4 * pow(t, 2);
 }
 
-fixp_s_ddd QuarticPolynomial::calc_third_derivative(double t) {
+fixp_s_ddd QuarticPolynomial::calc_third_derivative(float t) {
     return 6 * a3 + 24 * a4 * t;
 }
